Move 2tx-reconfig test that requires partitions
