+++ CMakeLists.txt
@@ -711,9 +711,7 @@ if(ENABLE_2TX_RECONFIG)
     NAME reconfiguration_test_cft_2tx
     PYTHON_SCRIPT ${CMAKE_SOURCE_DIR}/tests/reconfiguration.py
     CONSENSUS cft
-    ADDITIONAL_ARGS --ccf-version ${CCF_VERSION} --reconfiguration-type 2tx
-    LABEL partitions
-    CONFIGURATIONS partitions
+    ADDITIONAL_ARGS --ccf-version ${CCF_VERSION} --reconfiguration-type 2tx    
   )
 endif()
